requirements.txt
aiogram==2.24
aiohttp==3.8.3
aiosignal==1.3.1
APScheduler==3.9.1.post1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==2.1.1
frozenlist==1.3.3
idna==3.4
loguru==0.6.0
magic-filter==1.0.9
multidict==6.0.4
pyaes==1.6.1
pyasn1==0.4.8
python-dotenv==0.21.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
rsa==4.9
six==1.16.0
Telethon==1.26.1
tzdata==2022.7
tzlocal==4.2
yarl==1.8.2
